REFACTOR: Subclass server loading code from GraphLoader to ServerGraphLoader
Previously all code needed to do the following was in the GraphLoader class:
1. load cypher and parameters from files,
2. load connect to a Neo4j server instance, and
3. insert cypher data into that database

Now 1 is handled by the GraphLoader class. 2 and 3 are achieved by
ServerGraphLoader. This is useful as now GraphLoader can be subclassed again by
an EmbeddedGraphLoader class which will provide a embedded jython-friendly
interface to the cypher data. This will allow cymod to be used within Java
applicatons to provide data to an embedded Neo4j instance.

Using cymod within Java applications is a big help when using Neo4j and Cypher
in situations in which a Java application will have access to a JRE but no
Neo4j server will be available.